app/requirements-dev.txt
pytest==8.2.2
pytest-flask==1.3.0
bandit==1.7.9
safety==3.2.3
